Simplify - nn.Parameter depedency (#228)
* removes nn.Paramter dependency

* simple code update

+++ apps/Castor/modules/text_encoder.py
@@ -158,7 +158,7 @@ class Qwen2_5_VL(BaseTextEncoder):
                 return_tensors="pt",
                 max_length=self.text_seqlen,
                 truncation=True,
-            ).to(device=self.model.device, dtype=self.dtype)
+            ).to(device=self.model.device)
             
             if flops_meter is not None:
                 flops_meter.log_text_encoder_flops(inputs['input_ids'].shape)
